Extract logic into hook
